# foundry_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AcadAssist API",
    "description": "Backend API powering the AcadAssist AI Study Assistant via Microsoft Foundry.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/knowledge/search": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Search the student academic knowledge base",
        "description": "API boundary for Person 2 knowledge search. Filters by user_id and optionally course_id/subject_id.",
        "operationId": "search_knowledge_endpoint_api_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchKnowledgeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchKnowledgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/search_knowledge": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Search student academic knowledge base",
        "description": "Search academic materials using hybrid retrieval. Requires user_id and query.",
        "operationId": "search_knowledge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchKnowledgeParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchKnowledgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tools/get_progress": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Retrieve course completion progress",
        "description": "Retrieve current course completion percentage, reading milestone progress, and study streaks.",
        "operationId": "get_progress",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetProgressParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Progress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tools/create_study_plan": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Create an optimized study plan",
        "description": "Create an optimized daily study plan tailored to weak topics and available study hours.",
        "operationId": "create_study_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudyPlanParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Study Plan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/get_today_plan": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Retrieve today's study schedule",
        "description": "Retrieve the student's scheduled study sessions and practice tasks for today.",
        "operationId": "get_today_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTodayPlanParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Today Plan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/generate_weekly_report": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Generate a weekly learning summary",
        "description": "Generate a weekly learning summary including hours studied, concepts mastered, and streak status.",
        "operationId": "generate_weekly_report",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWeeklyReportParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate Weekly Report"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChunkResult": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id",
            "description": "UUID of the chunk"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "UUID of the source document"
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id",
            "description": "UUID of the course"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id",
            "description": "UUID of the subject"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Text content of the retrieved chunk"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title",
            "description": "Title of the source document"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title",
            "description": "Section or chapter title"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original filename of the document"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number",
            "description": "Page number if applicable"
          },
          "slide_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slide Number",
            "description": "Slide number if applicable"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Retrieval similarity score"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "content",
          "document_title",
          "filename",
          "score"
        ],
        "title": "ChunkResult",
        "description": "Single retrieved knowledge base chunk."
      },
      "CreateStudyPlanParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "description": "Start date (YYYY-MM-DD)"
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "description": "End date (YYYY-MM-DD)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "start_date",
          "end_date"
        ],
        "title": "CreateStudyPlanParams"
      },
...
      "GenerateWeeklyReportParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "week_start": {
            "type": "string",
            "title": "Week Start",
            "description": "Week start date (YYYY-MM-DD)"
          },
          "week_end": {
            "type": "string",
            "title": "Week End",
            "description": "Week end date (YYYY-MM-DD)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "week_start",
          "week_end"
        ],
        "title": "GenerateWeeklyReportParams"
      },
...
      "GetProgressParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetProgressParams"
      },
      "GetTodayPlanParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetTodayPlanParams"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SearchKnowledgeParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "query"
        ],
        "title": "SearchKnowledgeParams"
      },
      "SearchKnowledgeRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "UUID of the student user"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Search query string"
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id",
            "description": "Optional UUID filter for course"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id",
            "description": "Optional UUID filter for subject"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "Maximum number of chunks to return",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "query"
        ],
        "title": "SearchKnowledgeRequest",
        "description": "Request payload for searching the knowledge base."
      },
      "SearchKnowledgeResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ChunkResult"
            },
            "type": "array",
            "title": "Results",
            "description": "List of matching chunks"
          }
        },
        "type": "object",
        "title": "SearchKnowledgeResponse",
        "description": "Response payload containing retrieved knowledge chunks."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://lqs5ss0s-8000.inc1.devtunnels.ms"
    }
  ]
}